// commerceos/libs/apps/transactions/src/lib/shared/interfaces/detail.interface.ts
export interface ActionInterface {
  action: ActionTypeEnum;
  enabled: boolean;
  partialAllowed: boolean;
  isOptional: boolean;
  label?: string;
  description?: string;
  hidden?: boolean;
}

export interface ActionMapInterface {
  mainActions: UIActionInterface[];
  optionalActions: UIActionInterface[];
}

export interface MailActionInterface extends ActionInterface {
  mailEvent?: MailEventInterface;
}

export interface MailEventInterface {
  id: string;
}

export interface RefundItemsInterface {
  count: number;
  payment_item_id: string;
  item_uuid?: string;
  identifier?: string;
}
// ...
export interface ItemOptionInterface {
  name: string;
  value: string;
}

export interface ItemInterface {
  uuid: string;
  description: string;
  fixed_shipping_price: number;
  identifier: string;
  item_type: string;
  name: string;
  price: number;
  price_net: number;
  product_variant_uuid: string;
  quantity: number;
  shipping_price: number;
  shipping_settings_rate: number;
  shipping_settings_rate_type: string;
  shipping_type: string;
  thumbnail: string;
  updated_at: string;
  url: string;
  vat_rate: number;
  weight: number;
  created_at: string;
  options?: ItemOptionInterface[];
}

export interface RefundItemInterface {
  count: number;
  quantity?: number;
  item_uuid?: string;
  identifier?: string;
}

export interface RefundProductInterface {
  id: string;
  name: string;
  image: string;
  price: number;
  currency: string;
  quantity: number;
}
// ...
export interface ActionRequestRefundItemsInterface {
  identifier: string;
  count: number;
  price: number;
  quantity: number;
}

export interface ActionRequestUpdateDataInterface {
  deliveryFee?: string;
  productLine?: ItemInterface[];
}

export interface ActionRequestVerifyInterface {
  idNumber: string;
  applicant: number;
};

export interface ActionRequestInterface {
  amount?: number;
  dunningCosts?: number;
  customerId?: string;
  invoiceDate?: string;
  itemsRestocked?: boolean;
  reason?: string;
  delivery_fee?: number;
  refundCollectedBySepa?: boolean;
  payment_items?: ActionRequestRefundItemsInterface[];
  refundGoodsReturned?: boolean;
  refundInvoiceNumber?: string;
  status?: string;
  updateData?: ActionRequestUpdateDataInterface;
  identification?: ActionRequestVerifyInterface[]
}